extended the ruby integer class
started with some basics that i like to use when developing the API, but
i think i may extend this considerably more in the future. it's just
sometimes easier to work with integers rather than model objects.
especially when using the console. though if this slows things down with
otherwise unnecessary lookups, i might want to revisit this strategy.