email link back to invitation
will modify the invitation.show page such that user can log in then
select company and entity with which to link the invite to. submission
of this new form will invoke the investment.create action. first though
i need to have a way to link users to companies they manage.